requirements_for_apple_silicon.txt
absl-py==2.3.1
accelerate==1.1.1
addict==2.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
basicsr==1.4.2
blinker==1.9.0
bpy==4.0.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
coloredlogs==15.0.1
comm==0.2.3
ConfigArgParse==1.7
contourpy==1.3.2
cycler==0.12.1
Cython==3.1.2
dash==3.1.1
dataclasses-json==0.6.7
decorator==5.2.1
deepspeed==0.17.2
Deprecated==1.2.18
diffusers==0.30.0
einops==0.8.0
exceptiongroup==1.3.0
executing==2.2.0
facexlib==0.3.0
fast_simplification==0.1.11
fastapi==0.115.12
fastjsonschema==2.21.1
ffmpy==0.6.1
filelock==3.18.0
filterpy==1.4.5
Flask==3.1.1
flatbuffers==25.2.10
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.7.0
future==1.0.0
gfpgan==1.3.8
gradio==5.33.0
gradio_client==1.10.2
groovy==0.1.2
grpcio==1.74.0
h11==0.16.0
hf-xet==1.1.5
hjson==3.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.2
humanfriendly==10.0
idna==3.10
imageio==2.36.0
importlib_metadata==8.7.0
ipydatawidgets==4.3.5
ipython==8.37.0
ipywidgets==8.1.7
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
lazy_loader==0.4
lightning==2.5.2
lightning-utilities==0.15.0
llvmlite==0.44.0
lmdb==1.7.3
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
mypy_extensions==1.1.0
narwhals==1.48.1
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.1
numba==0.61.2
numpy==1.24.4
omegaconf==2.3.0
onnxruntime==1.16.3
open3d==0.18.0
opencv-python==4.10.0.84
opencv-python-headless==4.11.0.86
orjson==3.11.1
packaging==25.0
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
plotly==6.2.0
pooch==1.8.2
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==4.25.8
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pybind11==2.13.4
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
pygltflib==1.16.3
Pygments==2.19.2
PyMatting==1.1.14
pymeshlab==2025.7
pyparsing==3.2.3
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pythreejs==2.4.2
pytorch-lightning==2.5.2
pytz==2025.2
PyYAML==6.0.2
realesrgan==0.3.0
referencing==0.36.2
regex==2024.11.6
rembg==2.0.65
requests==2.32.4
retrying==1.4.1
rich==14.1.0
rpds-py==0.26.0
ruff==0.12.5
safehttpx==0.1.6
safetensors==0.4.4
scikit-image==0.24.0
scikit-learn==1.7.1
scipy==1.14.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.46.2
sympy==1.14.0
tb-nightly==2.18.0a20240726
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==1.0.19
tokenizers==0.20.3
tomli==2.2.1
tomlkit==0.13.3
torch==2.9.0.dev20250727
torchaudio==2.8.0.dev20250727
torchmetrics==1.8.0
torchvision==0.24.0.dev20250727
tqdm==4.66.5
traitlets==5.14.3
traittypes==0.2.1
transformers==4.46.0
trimesh==4.7.1
typer==0.16.0
typing-inspect==0.9.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.34.3
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wrapt==1.17.2
xatlas==0.0.11
yapf==0.43.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
